Move Messages and Address to individual files
