+++ crates/core/tedge_api/src/lib.rs
@@ -2,3 +2,5 @@
 
 pub mod plugins;
 pub mod errors;
+pub mod messages;
+pub mod address;
